API Pagination

+++ Post/api/views.py
@@ -1,12 +1,17 @@
 from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.generics import ListAPIView
+from rest_framework.authentication import TokenAuthentication
+from rest_framework.pagination import PageNumberPagination
 
 from Post.models import PostModel
 from Account.models import Account
 from .serializers import PostSerializers
 
 @api_view(['GET',])
+@permission_classes((IsAuthenticated, ))
 def api_detail_view(request, slug):
 	data = {}
 	try:
@@ -21,11 +26,16 @@ def api_detail_view(request, slug):
 		return Response(serializer.data)
 
 @api_view(['PUT', ])
+@permission_classes((IsAuthenticated, ))
 def api_update_view(request, slug):
 	try:
 		post_model = PostModel.objects.get(slug=slug)
 	except PostModel.DoesNotExist:
 		return Response(status=status.HTTP_404_NOT_FOUND)
+
+	user = request.user
+	if post_model.author != user:
+		return Response({'response': "You dont have permission to edit that"})
 
 	if request.method == 'PUT':
 		serializer = PostSerializers(post_model, data=request.data)
@@ -37,11 +47,16 @@ def api_update_view(request, slug):
 		return Response(serializer.errors, status=status.HTTP_404_BAD_REQUEST)
 
 @api_view(['DELETE'],)
+@permission_classes((IsAuthenticated, ))
 def api_delete_view(request, slug):
 	try:
 		post_model = PostModel.objects.get(slug=slug)
 	except PostModel.DoesNotExist:
 		return Response(status=status.HTTP_404_NOT_FOUND)
+
+	user = request.user
+	if post_model.author != user:
+		return Response({'response': "You dont have permission to edit that"})
 
 	if request.method == 'DELETE':
 		operations = post_model.delete()
@@ -53,9 +68,9 @@ def api_delete_view(request, slug):
 		return Response(data=data)
 
 @api_view(['POST'],)
+@permission_classes((IsAuthenticated, ))
 def api_create_view(request):
-	account = Account.objects.get(pk=1)
-	print(account)
+	account = request.user
 	if request.method == 'POST':
 		post_model = PostModel(author=account)
 		serializer = PostSerializers(post_model, data=request.data)
@@ -63,4 +78,12 @@ def api_create_view(request):
 			serializer.save()
 			return Response(serializer.data, status=status.HTTP_201_CREATED)
 		return Response(serializer.errors, staus=status.HTTP_404_BAD_REQUEST)
+
+class listPost(ListAPIView):
+	queryset = PostModel.objects.all()
+	serializer_class = PostSerializers
+	authentication_class = (TokenAuthentication)
+	permission_class = (IsAuthenticated)
+	pagination_class = PageNumberPagination
+
